+++ elmyra.ip.access.epo/CHANGES.rst
@@ -6,6 +6,7 @@ development
 ===========
 - middleware: OPS applicant family analytics enhancements
 - middleware: enhance query expression utility service for “applicant” field, e.g. “MAMMUT SPORTS GROUP AG”
+- middleware: fix number normalization of AT numbers
 
 0.70.1
 ======