+++ elmyra.ip.access.epo/CHANGES.rst
@@ -7,6 +7,7 @@ development
 - middleware: OPS applicant family analytics enhancements
 - middleware: enhance query expression utility service for “applicant” field, e.g. “MAMMUT SPORTS GROUP AG”
 - middleware: fix number normalization of AT numbers
+- ui: scroll to first result entry after paging
 
 0.70.1
 ======